+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/special/KorasiSword.kts
@@ -1,14 +1,18 @@
 package world.gregs.voidps.world.interact.entity.player.combat.melee.special
 
 import world.gregs.voidps.engine.client.update.task.viewport.Spiral
+import world.gregs.voidps.engine.entity.Registered
 import world.gregs.voidps.engine.entity.character.Character
+import world.gregs.voidps.engine.entity.character.contain.ItemChanged
 import world.gregs.voidps.engine.entity.character.npc.NPCs
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.Players
 import world.gregs.voidps.engine.entity.character.update.visual.setAnimation
 import world.gregs.voidps.engine.entity.character.update.visual.setGraphic
 import world.gregs.voidps.engine.entity.get
+import world.gregs.voidps.engine.entity.item.EquipSlot
 import world.gregs.voidps.engine.entity.item.Item
+import world.gregs.voidps.engine.entity.item.equipped
 import world.gregs.voidps.engine.entity.set
 import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
@@ -20,6 +24,36 @@ import world.gregs.voidps.world.interact.entity.player.combat.specialAttack
 import kotlin.random.Random
 
 fun isKorasisSword(item: Item?) = item != null && item.name == "korasis_sword"
+
+on<Registered>({ isKorasisSword(it.equipped(EquipSlot.Weapon)) }) { player: Player ->
+    updateWeapon(player, player.equipped(EquipSlot.Weapon))
+}
+
+on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && isKorasisSword(item) }) { player: Player ->
+    updateWeapon(player, item)
+}
+
+fun updateWeapon(player: Player, weapon: Item) {
+    player["attack_range"] = 1
+    player.weapon = weapon
+}
+
+on<CombatSwing>({ !swung() && isKorasisSword(it.weapon) }, Priority.LOW) { player: Player ->
+    player.setAnimation("korasis_sword_${
+        when (player.attackType) {
+            "chop" -> "chop"
+            else -> "slash"
+        }
+    }")
+    player.hit(target)
+    delay = 5
+}
+
+on<CombatHit>({ isKorasisSword(weapon) }) { player: Player ->
+    player.setAnimation("korasis_sword_block")
+}
+
+// Special attack
 
 val players: Players by inject()
 val npcs: NPCs by inject()
