--- game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/weapon/KorasisSword.kts
@@ -1,43 +1,0 @@
-package world.gregs.voidps.world.interact.entity.player.combat.melee.weapon
-
-import world.gregs.voidps.engine.entity.Registered
-import world.gregs.voidps.engine.entity.character.contain.ItemChanged
-import world.gregs.voidps.engine.entity.character.player.Player
-import world.gregs.voidps.engine.entity.character.update.visual.setAnimation
-import world.gregs.voidps.engine.entity.item.EquipSlot
-import world.gregs.voidps.engine.entity.item.Item
-import world.gregs.voidps.engine.entity.item.equipped
-import world.gregs.voidps.engine.entity.set
-import world.gregs.voidps.engine.event.Priority
-import world.gregs.voidps.engine.event.on
-import world.gregs.voidps.world.interact.entity.combat.*
-
-fun isKorasi(item: Item?) = item != null && item.name.startsWith("korasis_sword")
-
-on<Registered>({ isKorasi(it.equipped(EquipSlot.Weapon)) }) { player: Player ->
-    updateWeapon(player, player.equipped(EquipSlot.Weapon))
-}
-
-on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && isKorasi(item) }) { player: Player ->
-    updateWeapon(player, item)
-}
-
-fun updateWeapon(player: Player, weapon: Item) {
-    player["attack_range"] = 1
-    player.weapon = weapon
-}
-
-on<CombatSwing>({ !swung() && isKorasi(it.weapon) }, Priority.LOW) { player: Player ->
-    player.setAnimation("korasis_sword_${
-        when (player.attackType) {
-            "chop" -> "chop"
-            else -> "slash"
-        }
-    }")
-    player.hit(target)
-    delay = 5
-}
-
-on<CombatHit>({ isKorasi(weapon) }) { player: Player ->
-    player.setAnimation("korasis_sword_block")
-}
