+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/weapon/ZamorakianSpear.kts
@@ -12,13 +12,13 @@ import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.world.interact.entity.combat.*
 
-fun isZSpear(item: Item?) = item != null && item.name.startsWith("zamorakian_spear")
+fun isZamorakianSpear(item: Item?) = item != null && item.name.startsWith("zamorakian_spear")
 
-on<Registered>({ isZSpear(it.equipped(EquipSlot.Weapon)) }) { player: Player ->
+on<Registered>({ isZamorakianSpear(it.equipped(EquipSlot.Weapon)) }) { player: Player ->
     updateWeapon(player, player.equipped(EquipSlot.Weapon))
 }
 
-on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && isZSpear(item) }) { player: Player ->
+on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && isZamorakianSpear(item) }) { player: Player ->
     updateWeapon(player, item)
 }
 
@@ -27,7 +27,7 @@ fun updateWeapon(player: Player, weapon: Item) {
     player.weapon = weapon
 }
 
-on<CombatSwing>({ !swung() && isZSpear(it.weapon) }, Priority.LOW) { player: Player ->
+on<CombatSwing>({ !swung() && isZamorakianSpear(it.weapon) }, Priority.LOW) { player: Player ->
     player.setAnimation("zamorakian_spear_${
         when (player.attackType) {
             "block" -> "lunge"
@@ -38,6 +38,6 @@ on<CombatSwing>({ !swung() && isZSpear(it.weapon) }, Priority.LOW) { player: Pla
     delay = 4
 }
 
-on<CombatHit>({ isZSpear(weapon) }) { player: Player ->
+on<CombatHit>({ isZamorakianSpear(weapon) }) { player: Player ->
     player.setAnimation("zamorakian_spear_block")
 }
